oracle-doris-rows-exporter/requirements.txt
fastapi
uvicorn[standard]
gunicorn
fastapi-utilities
prometheus-client
oracledb
mysql-connector-python
